requirements.txt
# PyTorch & vision/audio libs
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Core dependencies
numpy
pandas
scikit-learn
tqdm
matplotlib
opencv-python-headless
scipy

# DeepFake Forensic Tools
imagehash
PyWavelets

# Streamlit App
streamlit
pillow

# Optional but recommended for safety
requests

# Concurrency & Testing
filelock
pytest
